requirements.txt
selenium==4.1.3
pytest==7.1.1
webdriver_manager==3.5.4
allure-pytest==2.9.45